Always log info about Contentful webhook receipts
This was useful for debugging our Contentful webhooks as they weren't working correctly. In this case the Contentful webhook didn't have the right content type set.

Instead of removing this logging, let's make it permenant so we can quickly review basic data on how the service is performing.

We don't want a unique Rollbar event to occur for every entry id so we remove it from the title and move it to the payload. This will mean there should only be a single rollbar item per environment with many occurences, much easier to manage.