Always precompile assets on CI so we can catch problems before deployments
Because pull requests run CI in RAILS_ENV=test the image builds and the tests pass fine. However, when the same happens for deployment with RAILS_ENV=production then it can fail to precomplile and we have to start a new pull request to fix it whilst the deployment pipeline is stuck.

It's not ideal to add every new critical variable to this file. Doing this ensures that the app can always load in order for running this rake command. No actual values are necessasry.