Prevent deployment failure when DOCKER_IMAGE contains a partial secret
Github will obfuscate any output that contains a secret IF that secret has been/is being used. We move the global `GITHUB_SECRETS_JSON` variable down into the deploy step where it's needed. We don't need all variables loaded into json for the build stage.

This doesn't get us all the way though since we had an echo in the build step:

```
echo "TF_VAR_docker_image=${{needs.build.outputs.tf_var_docker_image}}" >> $GITHUB_ENV
```

This line can set the `TF_VAR_DOCKER_IMAGE` to be a value that contains an obfuscating. Eg. with a GIT_SHA of 2387456***329487562. This doesn't exist when Terraform tries to deploy and it all goes badly.

Instead of echoing, we try passing the docker image to the script directly, rather than trying to set a new environment variable. The hope will be, because we don't echo, github doesn't change the docker image to include the asterisks.