chore: reorganise spec folders so they are ordered vertically
Now we have tasks in the hierarchy the folder structure doesn't follow the association hierachy.

We used to have:

- category
- sections
- steps

This was handy as when you add or work on the category, you scroll down to the next folder, do the work there and so on. This worked by accident due to the alphabetical sorting.

Since "Task" upsets this we add number prefixes to keep the folders organised logically:

- 1 category
- 2 sections
- 3 task
- 4 steps