Test suite waits for redis to be ready before running
Failurs - particularly on CI - have been randomly occuring recently. Sometimes the test suite runs before Redis is ready and a large part of the test suite fails due to failing Redis connections.

The debian `wait-for-it` package is used to wait for Redis to be ready _before_ running the tests. This follows this Docker guidance [1]. We give Redis a generous 60 seconds to start up.

Wait for it is only installed in the test stage of the Dockerfile so it should not be included in the deployed image.

[1] https://docs.docker.com/compose/startup-order/