Steps are presented to users in their Contentful Order
Doesn't seem limited to hidden questions but real case we found that
alerted us to this was for this type so this is the fixture we're using.

We use the same approach to ordering sections by using a `default_scope`
as `implicit_order_column` only works for finder methods like `first`
and `last`.