Cloning Contentful environments cache busts the category
Without this, the new environment will have no effect on the environment untill the currently cached content expires. This is 72 hours by default.

This change repurposes the Contentful webhook for cache busting and will target the default starting entry. This is the first entry that is requested from Contentful and if it's not found (thanks to this change), the service will go and ask Contentful for all the new records.

A workaround for this change would be to manually edit the category within Contentful. That would have the same affect. This change automates that, making it more likely to not be forgotten.

For testing of entries, the content designer can still edit individual entries as normal and they will be cache busted too.