Remove duplicate caching logic and explain the approach
As the comment says: GetEntry does the caching for any type of Contentful Entry.

We don't need to take explict steps as this is duplicating the effort.

Use a comment to explain to the reader that though no explicit "caching" looks to be happening, it's going on under the hood by requesting all the required data.