Cache warming will try to repopulate completely every 24 hours
The previous logic was probably not doing what we wanted.

It was going through every entry and asking GetEntry to find it for us. This service has built in caching so it would check the cache first. Effectively this would only fill in the gaps where entries were missing.

The only time an entry would be missing would be if it was a new entry added to Contentful.

The caching is in 24 hour blocks with a default cache value of 72 hours. Every night when the cache warming task comes along, there will always be contents in our cache and no opportunity to get a new clean set.

This change changes approach. We backup everything in our redis cache and fetch everything a fresh. This makes sure we are picking up any new changes made within the last 24 hours.

If it fails, we put all the old cache items back and extend them for another 24 hours. This can go on indefinitely.