Refactor test suite to separate how to arrive at the first question
* previously the spec helper was (helpfully) taking us from category to question. This is great but more logic has since been layered in which makes this silver bullet hard to work with:

1. for tasks that have one step, we behave very differently. The user is taken past the task page and on answering straight back to the journey page!
2. tasks with multiple steps route the user to a task page where they can click a task to start

This change breaks the helper into 2 to help reflect this logic. The only common behaviour is to go from category to the first section.

With this change in we will be able to make add a new third bit of logic: when a task has not been startd, we redirect them to the first question straight away and then back to the task page.