Refactor creating a journey to inject the category ID
Previously, we were providing a category name as a parameter when
creating a new journey, but we were pulling the Contentful entry ID of
the category from the environment within the CreateJourney service.

This change moves to injecting the entry ID as a parameter at create
time, and then pulling the title of the category from this entry.

This makes the intent of the method clearer, as well as making it easier
to add multiple spend categories in future.