Rewrite SeedSchool task so that it is less memory intensive
The main issue seemed to involve the Dry::Transformer module
processing the data and not freeing up the application memory when
it was done, so the number of objects in memory quickly got out
of hand. I forced the Garbage Collector to run earlier to free
the memory up which vastly reduced the memory usage. I also
changed the seed_schools task to seperate out the records that
needed to be updated from the ones that needed to be created and
ran them as two seperate methods, processing them in batches of
2000 and wrapping it up in a single transaction.