Remove broken attempt to find hostname on Linux
This snippet didn't work on Linux, at least for me: instead of returning
a hostname, the shell snippet returned multiple lines which could not be
resolved.

Hypothesis: the default behaviour of not specifying the Capybara host
will result in a sensible default of localhost. By setting server_host
iff the environment variable is set, we can fall back to Capybara's
reasonable default.