Add Git guidelines
The convention previously used, of squashing commits into large
sprawling collections of disparate changes, makes it hard to understand
the intention and decisions behind those changes. This is a proposal to
change the convention so that we have a more useful history in the
future.

This draws heavily on the GOV.UK Git style guide:
https://github.com/alphagov/styleguides/blob/master/git.md