Simplify agent controllers
Make AgentForm a subclass. This makes the implementation much simpler,
as it now adds only validations and role assignment on save.

As this is the only place that the assign_roles in RoleAssignable was
used, it's simpler to implement it directly in this class.

By removing this behaviour from the controllers, they become generic
enough that we don't have to feel bad about the duplication, which
lets us stop misusing a Concern to share controller functionality.

The fact that these share templates is still a wart, as is the fact that
there are two near-identical controllers.