feat(cms): Remove archived establishments from selection options
Archived organisations and establishment groups are those that are in our databases but are no longer in the GIAS data download for various reasons, such as a school never opening. We do not want Procurement Operations or school users to be able to select archived schools from any of the dropdowns or school pickers as there should be an alternate correct establishment to choose from.