Add tasklist tab to case
A tasklist tab is needed for the school collab project to show procops what tasks there are to complete for an evaluation. This is only for level 4 and 5 cases. At the moment, it is showing tasks but it is not functional and pages will be added one by one. It is behind a feature flag sc_tasklist_tab so the page is hidden until fully ready for procops.