Fix error when editing templates with attachments
In the great email sync refactor of October 2023, a single commit
988bc05 that changed 3700 lines of code, the
Support::Emails::Attachments class was removed.

However, one reference to it still remained. When editing a template
with pre-existing attachments, this would attempt to call a method on
the non-existent class, leading to an exception and not updating the
template.

The blob_attachments field is used to retain attachments across edits by
holding information about already-uploaded files associated with the
template. On save, the attachments are re-linked by dereferencing them
from the bag of JSON describing them. This method was lost in the
deleted class, although similar behaviour exists in
Email::Draft::Attachable. This is used as a reference to recreate
equivalent functionality here and to make template editing behave as
intended.